+++ src/util/Memory.h
@@ -4,27 +4,27 @@
 namespace Util {
 
 template<typename T>
-void Copy(T* from, T* to, int amount)
+void Copy(T* from, T* to, size_t amount)
 {
-  for (int i = 0; i < amount; ++i)
+  for (size_t i = 0; i < amount; ++i)
   {
     to[i] = from[i];
   }
 }
 
 template<typename T>
-void MoveData(T* from, T* to, int amount)
+void MoveData(T* from, T* to, size_t amount)
 {
-  for (int i = 0; i < amount; ++i)
+  for (size_t i = 0; i < amount; ++i)
   {
     to[i] = Move(from[i]);
   }
 }
 
 template<typename T>
-void Fill(T* data, const T& value, int amount)
+void Fill(T* data, const T& value, size_t amount)
 {
-  for (int i = 0; i < amount; ++i)
+  for (size_t i = 0; i < amount; ++i)
   {
     data[i] = value;
   }
